# apps/api/openapi.spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/{id}/keys": {
      "post": {
        "operationId": "KeysController_create",
        "summary": "Create a new API key",
        "description": "Create a new API Key",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of the project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateKeyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Tried creating a 'live' key without verified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "404": {
            "description": "No project found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Keys"
        ]
      },
      "get": {
        "operationId": "KeysController_findAll",
        "summary": "Fetch all keys related to project",
        "description": "Fetch all keys related to project",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of the project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchAllKeysResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Keys"
        ]
      }
    },
    "/projects/{id}/keys/{keyId}": {
      "get": {
        "operationId": "KeysController_findOne",
        "summary": "Fetch a key by id",
        "description": "Fetch a key by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of the project",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyId",
            "required": true,
            "in": "path",
            "description": "Id of the key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchKeyResponse"
                }
              }
            }
          },
          "404": {
            "description": "No key found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Keys"
        ]
      }
    },
    "/projects/{id}/keys/verify": {
      "post": {
        "operationId": "KeysController_verifyKey",
        "summary": "Verify a key",
        "description": "Verify a key",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of the project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyKeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyKeyEntity"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "404": {
            "description": "Either this API Key does not belong to this project or is incorrect.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Keys"
        ]
      }
    },
...
  },
  "info": {
    "title": "LogsTrap",
    "description": "A functional logging solution",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header",
        "x-tokenName": "Token"
      },
      "Api-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "GenericErrorEntity": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "ProjectMode": {
        "type": "string",
        "enum": [
          "test",
          "live"
        ]
      },
...
      "CreateKeyDto": {
        "type": "object",
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/ProjectMode"
          },
          "description": {
            "type": "string",
            "description": "Description for the key"
          },
          "name": {
            "type": "string",
            "description": "Name of the API Key, will be used for searching and sorting"
          }
        },
        "required": [
          "mode",
          "name"
        ]
      },
      "CreateKeyResponse": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "API Key, must be stored as it can't be seen again"
          },
          "id": {
            "type": "string",
            "description": "Id of key in db"
          }
        },
        "required": [
          "key",
          "id"
        ]
      },
      "Key": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ProjectMode"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "createdAt",
          "mode",
          "name"
        ]
      },
      "FetchAllKeysResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Key"
            }
          },
          "totalItems": {
            "type": "number"
          },
          "itemsPerQuery": {
            "type": "number",
            "default": 6
          }
        },
        "required": [
          "items",
          "totalItems",
          "itemsPerQuery"
        ]
      },
      "FetchKeyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ProjectMode"
          },
          "name": {
            "type": "string"
          },
          "apiRequestsCount": {
            "type": "number"
          },
          "applicationLogsCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "projectId",
          "createdAt",
          "mode",
          "name",
          "apiRequestsCount",
          "applicationLogsCount"
        ]
      },
      "VerifyKeyDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          }
        },
        "required": [
          "key"
        ]
      },
      "VerifyKeyEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the key"
          },
          "name": {
            "type": "string",
            "description": "The name of the key"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
    }
  }
}